Renaming `elementConfigurationSections` directory to `elementConfiguration`.
