+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfiguration/Metric.tsx
