+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfiguration/VisualizationConfigurationOptions.tsx
@@ -18,7 +18,7 @@ import * as React from 'react';
 import { Field, getIn, useFormikContext } from 'formik';
 import { ConfigurationField } from 'views/types';
 
-import BooleanField from 'views/components/aggregationwizard/elementConfigurationSections/configurationFields/BooleanField';
+import BooleanField from 'views/components/aggregationwizard/elementConfiguration/configurationFields/BooleanField';
 import { VisualizationConfigFormValues } from 'views/components/aggregationwizard/WidgetConfigForm';
 import { HoverForHelp } from 'components/common';
